Now accurately finds all points in the letters
I tweaked the algorithm to find the change in brightness of colors 2
pixels away, and if it is outside a certain threshold, it will be
checked. Then it fills in similar colors around it until it is a
completed letter. Added a slider to change the fuzziness for finding
similar colors.